requirements.txt
aiohttp
google-generativeai
loguru
pymongo
twilio
python-dotenv
schedule
# Install browser-use version that doesn't depend on langchain-google-genai
browser-use
openai
langchain-openai
# Remove crawl4ai, as we're no longer using it
# crawl4ai